pyproject.toml
[project]
name = "dry-check"
version = "0.1.1"
description = "Notify when it stops raining"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "PyYAML",
]
 
